refactor: DRY section sorting logic into shared utility
Extract sortSectionsByLinkedList to src/common/utils/sections.ts
to avoid duplication between frontend (workspaceFiltering.ts) and
backend (projectService.ts).
