+++ src/browser/utils/ui/workspaceFiltering.ts
@@ -1,6 +1,9 @@
 import type { FrontendWorkspaceMetadata } from "@/common/types/workspace";
 import type { ProjectConfig, SectionConfig } from "@/common/types/project";
 
+// Re-export shared section sorting utility
+export { sortSectionsByLinkedList } from "@/common/utils/sections";
+
 function flattenWorkspaceTree(
   workspaces: FrontendWorkspaceMetadata[]
 ): FrontendWorkspaceMetadata[] {
@@ -357,45 +360,6 @@ export function getVisibleWorkspaces(
   }
 
   return visible;
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Section ordering and grouping
-// ─────────────────────────────────────────────────────────────────────────────
-
-/**
- * Sort sections by their linked-list order (nextId pointers).
- * Matches the backend's sortSectionsByLinkedList logic.
- */
-export function sortSectionsByLinkedList(sections: SectionConfig[]): SectionConfig[] {
-  if (sections.length === 0) return [];
-
-  const byId = new Map(sections.map((s) => [s.id, s]));
-  // Find head: section not referenced by any other section's nextId
-  const referencedIds = new Set(sections.map((s) => s.nextId).filter(Boolean));
-  const heads = sections.filter((s) => !referencedIds.has(s.id));
-
-  // If no clear head (cycle or empty), fall back to first section
-  const head = heads[0] ?? sections[0];
-
-  const sorted: SectionConfig[] = [];
-  const visited = new Set<string>();
-  let current: SectionConfig | undefined = head;
-
-  while (current && !visited.has(current.id)) {
-    visited.add(current.id);
-    sorted.push(current);
-    current = current.nextId ? byId.get(current.nextId) : undefined;
-  }
-
-  // Append orphaned sections (not in linked list)
-  for (const s of sections) {
-    if (!visited.has(s.id)) {
-      sorted.push(s);
-    }
-  }
-
-  return sorted;
 }
 
 // ─────────────────────────────────────────────────────────────────────────────
